test(client): testing change track order failure toast
