+++ packages/client/jest.setup.tsx
@@ -323,5 +323,5 @@ beforeEach(() => {
     });
 });
 
-jest.spyOn(console, 'warn').mockImplementation();
-jest.spyOn(console, 'error').mockImplementation();
+// jest.spyOn(console, 'warn').mockImplementation();
+// jest.spyOn(console, 'error').mockImplementation();
