+++ packages/client/screens/MusicPlaylistEditorRoomScreen/__tests__/TracksListChangeTrackOrder.test.tsx
@@ -1,35 +1,27 @@
 import {
     AllClientToServerEvents,
     MpeChangeTrackOrderOperationToApply,
-    MpeWorkflowState,
 } from '@musicroom/types';
-import { datatype } from 'faker';
 import Toast from 'react-native-toast-message';
-import { serverSocket } from '../../../services/websockets';
-import { db, generateTrackMetadata } from '../../../tests/data';
+import { db } from '../../../tests/data';
 import {
-    renderApp,
     fireEvent,
     waitFor,
     within,
     waitForElementToBeRemoved,
     render,
-} from '../../../tests/tests-utils';
-import {
-    createMpeRoom,
-    DefinedStateRef,
     toTrackCardContainerTestID,
-} from './TracksListEditing.test';
+} from '../../../tests/tests-utils';
+import { createMpeRoom, DefinedStateRef } from '../../../tests/tests-mpe-utils';
+import { serverSocket } from '../../../services/websockets';
 
 function extractTrackIDFromCardContainerTestID(testID: string): string {
     return testID.replace('-track-card-container', '');
 }
 
 async function playlistUIHasUnfreezed(screen: ReturnType<typeof render>) {
-    await waitFor(() => {
-        const [addTrackButton] = screen.getAllByText(/add.*track/i);
-
-        expect(addTrackButton).not.toBeDisabled();
+    return await waitFor(() => {
+        expect(screen.getByText(/add.*track/i)).not.toBeDisabled();
     });
 }
 
@@ -108,11 +100,13 @@ async function addTrack({
  * Will emit corresponding client socket event to perform a change track order operation
  * Will also check that track has been moved
  * Warning: don't forget to init the serverSocket handlers
+ * ShouldFail is an optionnal parameter that allow to fake a failure upper a valid operation
  */
 async function changeTrackOrder({
     screen,
     state,
     trackToMove: { operationToApply, fromIndex },
+    shouldFail,
 }: {
     state: DefinedStateRef;
     screen: ReturnType<typeof render>;
@@ -120,6 +114,7 @@ async function changeTrackOrder({
         fromIndex: number;
         operationToApply: MpeChangeTrackOrderOperationToApply;
     };
+    shouldFail?: boolean;
 }): Promise<void> {
     const destIndex =
         fromIndex +
@@ -137,9 +132,7 @@ async function changeTrackOrder({
         extractTrackIDFromCardContainerTestID(testID),
     );
 
-    await waitFor(async () => {
-        await playlistUIHasUnfreezed(screen);
-    });
+    await playlistUIHasUnfreezed(screen);
 
     expect(fromIndex < state.value.tracks.length).toBeTruthy();
 
@@ -172,23 +165,27 @@ async function changeTrackOrder({
         fireEvent.press(moveUpTrackButton);
     }
 
-    await waitFor(async () => {
-        await playlistUIHasUnfreezed(screen);
-        const trackCardElements =
-            screen.getAllByTestId(/track-card-container/i);
+    if (!shouldFail) {
+        await waitFor(() => {
+            const trackCardElements =
+                screen.getAllByTestId(/track-card-container/i);
+            expect(trackCardElements.length).toBe(state.value.tracks.length);
 
-        /**
-         * Tracks have been swapped
-         */
-        expect(trackCardElements[fromIndex]).toHaveProp(
-            'testID',
-            toTrackCardContainerTestID(tracksIDs[destIndex]),
-        );
-        expect(trackCardElements[destIndex]).toHaveProp(
-            'testID',
-            toTrackCardContainerTestID(tracksIDs[fromIndex]),
-        );
-    });
+            /**
+             * Tracks have been swapped
+             */
+            expect(trackCardElements[fromIndex]).toHaveProp(
+                'testID',
+                toTrackCardContainerTestID(tracksIDs[destIndex]),
+            );
+            expect(trackCardElements[destIndex]).toHaveProp(
+                'testID',
+                toTrackCardContainerTestID(tracksIDs[fromIndex]),
+            );
+        });
+
+        await playlistUIHasUnfreezed(screen);
+    }
 }
 
 test('Move track', async () => {
@@ -243,7 +240,6 @@ test('Move track', async () => {
         trackToAdd: fakeTracks[0],
     });
 
-    console.log('FIRST TRACK', state.value);
     await changeTrackOrder({
         screen,
         state,
@@ -253,7 +249,14 @@ test('Move track', async () => {
         },
     });
 
-    console.log('SECOND TRACK', state.value);
+    //First time is during add track step
+    await waitFor(() => {
+        expect(Toast.show).toHaveBeenNthCalledWith(2, {
+            type: 'success',
+            text1: expect.any(String),
+        });
+    });
+
     await changeTrackOrder({
         screen,
         state,
@@ -261,5 +264,59 @@ test('Move track', async () => {
             fromIndex: 1,
             operationToApply: MpeChangeTrackOrderOperationToApply.Values.UP,
         },
+    });
+
+    await waitFor(() => {
+        expect(Toast.show).toHaveBeenNthCalledWith(3, {
+            type: 'success',
+            text1: expect.any(String),
+        });
+    });
+});
+
+//hacking my way out to display a change track order error toast
+test('after receiving a MPE_CHANGE_TRACK_ORDER_FAIL_CALLBACK it should display a failure toast', async () => {
+    const { screen, state } = await createMpeRoom();
+    const fakeTracks = [
+        db.searchableTracks.create(),
+        db.searchableTracks.create(),
+    ];
+
+    //Init serverSocket listeners
+    serverSocket.on('MPE_CHANGE_TRACK_ORDER_DOWN', ({ roomID }) => {
+        serverSocket.emit('MPE_CHANGE_TRACK_ORDER_FAIL_CALLBACK', {
+            roomID,
+        });
+    });
+
+    serverSocket.on('MPE_CHANGE_TRACK_ORDER_UP', ({ roomID }) => {
+        serverSocket.emit('MPE_CHANGE_TRACK_ORDER_FAIL_CALLBACK', {
+            roomID,
+        });
+    });
+    ///
+
+    await addTrack({
+        screen,
+        state,
+        trackToAdd: fakeTracks[0],
+    });
+
+    await changeTrackOrder({
+        screen,
+        state,
+        shouldFail: true,
+        trackToMove: {
+            fromIndex: 0,
+            operationToApply: MpeChangeTrackOrderOperationToApply.Values.DOWN,
+        },
+    });
+
+    //First time is during add track step
+    await waitFor(() => {
+        expect(Toast.show).toHaveBeenNthCalledWith(2, {
+            type: 'error',
+            text1: expect.any(String),
+        });
     });
 });
